+++ src/core/runtime.ts
@@ -65,6 +65,9 @@ class Runtime {
             case 'parse':
                 this.parseHTML(data.body, data.requestUid);
                 break;
+            case 'mount-inline-scripts':
+                this.handleInlineScripts(data.selector);
+                break;
             default:
                 return;
         }
@@ -123,6 +126,30 @@ class Runtime {
                 break;
             default:
                 return;
+        }
+    }
+
+    /**
+     * Looks through the new HTML for any inline scripts and attempts to append them to the documents head.
+     */
+    private handleInlineScripts(selector: string): void {
+        const el = document.body.querySelector(selector);
+        if (el) {
+            el.querySelectorAll('script').forEach(script => {
+                const newScript = document.createElement('script');
+                if (script?.src || script?.id || script.getAttribute('pjax-script-id')) {
+                    let scriptSelector = 'script';
+                    scriptSelector += `[src="${script?.src}"]` || `#${script?.id}` || `[pjax-script-id="${script.getAttribute('pjax-script-id')}"]`;
+                    const existingScript = document.head.querySelector(scriptSelector);
+                    if (existingScript) {
+                        existingScript.remove();
+                    }
+                    newScript.src = script.src;
+                } else {
+                    newScript.innerHTML = script.innerHTML;
+                }
+                document.head.appendChild(newScript);
+            });
         }
     }